fix: Fix alignment error.
ref: https://github.com/go-ole/go-ole/issues/147

+++ safearray/safearray.go
@@ -33,6 +33,6 @@ type SafeArray struct {
 	FeaturesFlag uint16
 	ElementsSize uint32
 	LocksAmount  uint32
-	Data         uint32
+	Data         uintptr
 	Bounds       [16]byte
 }
